pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thermographic-report-builder"
version = "0.1.0"
description = "Production-ready thermographic report builder for solar panel inspection"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Aisol", email = "dev@aisol.cloud"}
]

dependencies = [
    # Core dependencies
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # AWS
    "boto3>=1.34.0",

    # Image processing
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "rasterio>=1.3.9",

    # Geospatial
    "shapely>=2.0.0",
    "pyproj>=3.6.0",

    # Report generation
    "pylatex>=1.4.0",
    "reportlab>=4.0.0",

    # Optional: DXF support
    "ezdxf>=1.1.0",

    # Utilities
    "python-json-logger>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "ipython>=8.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/thermographic_report_builder --cov-report=html --cov-report=term"
